update src for playstore and app store imgs, move qrcode and playstore btns into separate component
